Move child_process and events to llrt_modules
