+++ llrt_core/src/lib.rs
@@ -4,7 +4,6 @@
 
 #[macro_use]
 mod macros;
-mod bytearray_buffer;
 mod bytecode;
 // #[cfg(not(feature = "lambda"))]
 pub mod compiler;
@@ -18,7 +17,9 @@ pub mod modules;
 pub mod number;
 pub mod runtime_client;
 mod security;
-mod stream;
+mod stream {
+    pub use llrt_modules::stream::*;
+}
 mod test_utils;
 pub mod utils;
 pub mod vm;
