+++ llrt_core/src/macros.rs
@@ -14,12 +14,3 @@ macro_rules! iterable_enum {
         iterable_enum!(, $name, $($member),*)
     };
 }
-
-macro_rules! impl_stream_events {
-
-    ($($struct:ident),*) => {
-        $(
-            impl<'js> $crate::stream::SteamEvents<'js> for $struct<'js> {}
-        )*
-    };
-}
