+++ llrt_core/src/modules/mod.rs
@@ -1,6 +1,6 @@
 pub use llrt_modules::{buffer, fs, os, path};
 
-pub mod child_process;
+// pub mod child_process;
 pub mod console;
 pub mod crypto;
 pub mod encoding;
