+++ llrt_modules/Cargo.toml
@@ -6,9 +6,11 @@ license-file = "LICENSE"
 
 [features]
 default = ["all"]
-all = ["buffer", "fs", "os", "path", "process"]
+all = ["buffer", "child-process", "events", "fs", "os", "path", "process"]
 
 buffer = ["llrt_utils/encoding"]
+child-process = ["events", "__stream", "itoa"]
+events = []
 fs = ["tokio/fs", "llrt_utils/fs", "ring", "buffer", "path"]
 os = [
   "process",
@@ -21,17 +23,23 @@ os = [
 path = []
 process = []
 
+__bytearray-buffer = ["tokio/sync"]
+__stream = ["buffer", "__bytearray-buffer"]
+
 [dependencies]
 either = "1"
+itoa = { version = "1", optional = true }
 once_cell = { version = "1", optional = true }
 llrt_utils = { path = "../llrt_utils", default-features = false }
 rquickjs = { version = "0.6", features = [
+  "array-buffer",
   "either",
   "macro",
   "futures",
 ], default-features = false }
 ring = { version = "0.17", optional = true }
 tokio = { version = "1", features = ["rt", "io-util"] }
+tracing = "0.1"
 
 [target.'cfg(unix)'.dependencies]
 libc = { version = "0.2", optional = true }
