+++ llrt_modules/src/lib.rs
@@ -1,10 +1,13 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
+#![allow(clippy::new_without_default)]
 pub use self::module_info::ModuleInfo;
 pub use self::modules::*;
 
 mod module_info;
 mod modules;
+#[cfg(feature = "__stream")]
+pub mod stream;
 #[cfg(test)]
 mod test;
 mod utils;
