+++ llrt_modules/src/utils/mod.rs
@@ -1,3 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 pub mod array_buffer;
+#[cfg(feature = "__bytearray-buffer")]
+pub mod bytearray_buffer;
